requirements.txt
# Core dependencies
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.10.6
pandas==2.2.2

# Storage
boto3==1.37.17
botocore==1.37.17

# Analytics (Optional: Use if you want Hume AI analytics)
hume==0.7.0

# LLM integration (Optional: Use if you want LLM for insights)
anthropic==0.8.1

# PDF generation (Optional: Use if you want to generate PDF reports)
reportlab==4.0.7
PyPDF2==3.0.1

# Email
yagmail==0.15.293
imaplib2==3.6
email-validator==2.2.0
google-api-python-client==2.107.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0

# Database
SQLAlchemy==2.0.39

# Scheduling
schedule==1.2.1
apscheduler==3.11.0

# CLI utilities
tabulate==0.9.0

# Utilities
python-dateutil==2.8.2
tqdm==4.67.1
pytz==2025.1